Remove old translation_service.rs file and update test structure
Short description: After restructuring the translation module, remove the old file and update test references to ensure code clarity

Model: claude-3.7-sonnet-thinking

Prompt: There was a lot of trouble around the imports fixing, make sure there won't be any confusion again about translation and translation_service.

Thought Process:
To prevent confusion between the old translation_service.rs file and the new modular translation module structure, I've completely removed the old file which was likely causing import ambiguity. I've also renamed the test file from translation_service_tests.rs to translation_tests.rs and updated the imports in tests/main.rs to match. The old file was redundant since we now have a proper modular structure with core.rs, batch.rs, formatting.rs, and cache.rs files in the translation directory. I also added documentation in the translation_tests.rs file to clarify the new module structure. This should make the codebase more maintainable and prevent future confusion about imports.

Discussion:
There was initially confusion because both the old translation_service.rs file and the new translation module structure existed at the same time, causing ambiguous imports. The solution was to fully remove the old file and ensure all tests properly reference the new module structure. Tests are now passing successfully.
