--- src/translation_service.rs
@@ -1,2150 +1,0 @@
-use anyhow::{Result, Context, anyhow};
-use log::{error, warn, info};
-use std::time::{Duration, Instant};
-use url::Url;
-use std::sync::Arc;
-use regex::Regex;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Mutex as StdMutex;
-use tokio::sync::Semaphore;
-use std::fmt;
-use std::str::FromStr;
-use futures::stream::{self, StreamExt};
-
-use crate::app_config::{TranslationConfig, TranslationProvider as ConfigTranslationProvider};
-use crate::subtitle_processor::SubtitleEntry;
-use crate::providers::ollama::{Ollama, GenerationRequest};
-use crate::providers::openai::{OpenAI, OpenAIRequest};
-use crate::providers::anthropic::{Anthropic, AnthropicRequest};
-
-// @module: Translation service for subtitle content
-
-// @struct: Token usage statistics
-#[derive(Clone)]
-pub struct TokenUsageStats {
-    // @field: Number of prompt tokens
-    pub prompt_tokens: u64,
-    
-    // @field: Number of completion tokens
-    pub completion_tokens: u64,
-    
-    // @field: Total number of tokens
-    pub total_tokens: u64,
-    
-    // @field: Start time of token tracking
-    pub start_time: Instant,
-    
-    // @field: Total time spent on API requests
-    pub api_duration: Duration,
-    
-    // @field: Provider name
-    pub provider: String,
-    
-    // @field: Model name
-    pub model: String,
-}
-
-impl TokenUsageStats {
-    // @creates: New empty token usage stats
-    pub fn new() -> Self {
-        Self {
-            prompt_tokens: 0,
-            completion_tokens: 0,
-            total_tokens: 0,
-            start_time: Instant::now(),
-            api_duration: Duration::from_secs(0),
-            provider: String::new(),
-            model: String::new(),
-        }
-    }
-    
-    // @updates: Add token usage numbers
-    pub fn add_token_usage(&mut self, prompt_tokens: Option<u64>, completion_tokens: Option<u64>) {
-        if let Some(pt) = prompt_tokens {
-            self.prompt_tokens += pt;
-            self.total_tokens += pt;
-        }
-        
-        if let Some(ct) = completion_tokens {
-            self.completion_tokens += ct;
-            self.total_tokens += ct;
-        }
-    }
-    
-    // @updates: Add API request duration
-    pub fn add_request_duration(&mut self, duration: Duration) {
-        self.api_duration += duration;
-    }
-    
-    // @creates: New token usage stats with provider info
-    pub fn with_provider_info(provider: String, model: String) -> Self {
-        Self {
-            prompt_tokens: 0,
-            completion_tokens: 0,
-            total_tokens: 0,
-            start_time: Instant::now(),
-            api_duration: Duration::from_secs(0),
-            provider,
-            model,
-        }
-    }
-    
-    // @returns: Tokens per minute rate
-    pub fn tokens_per_minute(&self) -> f64 {
-        // Use the API duration for rate calculation, with fallback to elapsed time
-        let duration_minutes = if self.api_duration.as_secs_f64() > 0.0 {
-            self.api_duration.as_secs_f64() / 60.0
-        } else {
-            self.start_time.elapsed().as_secs_f64() / 60.0
-        };
-        
-        // Calculate the rate, handling division by zero
-        if duration_minutes > 0.0 {
-            self.total_tokens as f64 / duration_minutes
-        } else {
-            0.0
-        }
-    }
-    
-    // @returns: Summary of token usage as a string
-    pub fn summary(&self) -> String {
-        // Calculate tokens per minute
-        let tokens_per_min = self.tokens_per_minute();
-        
-        // Format the API duration if available
-        let api_time = if self.api_duration.as_secs_f64() > 0.0 {
-            format!("in {:.1}s of API time", self.api_duration.as_secs_f64())
-        } else {
-            "".to_string()
-        };
-        
-        format!(
-            "Token usage: {} total ({} prompt, {} completion) {} at {:.0} tokens/min", 
-            self.total_tokens, 
-            self.prompt_tokens, 
-            self.completion_tokens,
-            api_time,
-            tokens_per_min
-        )
-    }
-}
-
-// @parses: Endpoint string into host and port
-// @returns: Tuple of (host, port)
-fn parse_endpoint(endpoint: &str) -> Result<(String, u16)> {
-    // If it doesn't start with http/https, assume it's just host:port
-    let url_str = if !endpoint.starts_with("http://") && !endpoint.starts_with("https://") {
-        format!("http://{}", endpoint)
-    } else {
-        endpoint.to_string()
-    };
-    
-    let url = Url::parse(&url_str)
-        .context(format!("Failed to parse endpoint URL: {}", endpoint))?;
-        
-    let host = format!(
-        "{}://{}",
-        url.scheme(),
-        url.host_str().unwrap_or("localhost")
-    );
-    
-    let port = url.port().unwrap_or(11434);
-    
-    Ok((host, port))
-}
-
-// @enum: Available translation provider implementations
-enum TranslationProviderImpl {    
-    // @variant: Ollama LLM service
-    Ollama {
-        // @field: Client instance
-        client: Ollama,
-    },
-    
-    // @variant: OpenAI API service
-    OpenAI {
-        // @field: Client instance
-        client: OpenAI,
-    },
-    
-    // @variant: Anthropic API service
-    Anthropic {
-        // @field: Client instance
-        client: Anthropic,
-    },
-}
-
-// @struct: Translation service
-pub struct TranslationService {
-    // @field: Provider implementation
-    provider: TranslationProviderImpl,
-    
-    // @field: Configuration
-    config: TranslationConfig,
-}
-
-/// Additional parsing for OpenAI responses
-/// OpenAI responses often follow certain patterns that we can try to detect
-fn extract_openai_response(response: &str) -> Option<String> {
-    // Check for itemized translations (common in OpenAI responses)
-    // Pattern: numbered entries like "1. Text", "2. Text", etc.
-    let numbered_regex = Regex::new(r"(?m)^\s*\d+\.\s*(.+)$").ok()?;
-    let matches: Vec<String> = numbered_regex
-        .captures_iter(response)
-        .filter_map(|cap| cap.get(1).map(|m| m.as_str().trim().to_string()))
-        .collect();
-    
-    if !matches.is_empty() {
-        return Some(matches.join("\n"));
-    }
-    
-    // Check for bullet point translations
-    // Pattern: bullet points like "• Text", "- Text", etc.
-    let bullet_regex = Regex::new(r"(?m)^\s*[•\-*]\s*(.+)$").ok()?;
-    let matches: Vec<String> = bullet_regex
-        .captures_iter(response)
-        .filter_map(|cap| cap.get(1).map(|m| m.as_str().trim().to_string()))
-        .collect();
-    
-    if !matches.is_empty() {
-        return Some(matches.join("\n"));
-    }
-    
-    // If no specific pattern is found, split by newlines and remove empty lines
-    let lines: Vec<&str> = response
-        .lines()
-        .map(|s| s.trim())
-        .filter(|s| !s.is_empty())
-        .collect();
-    
-    if !lines.is_empty() {
-        // Detect and remove common OpenAI response patterns at the start/end
-        let filtered_lines: Vec<&str> = lines
-            .into_iter()
-            .filter(|line| {
-                // Skip commentary lines that are not translations
-                !line.starts_with("Here's the translation") &&
-                !line.starts_with("I've translated") &&
-                !line.starts_with("Translation:") &&
-                !line.contains("from English to") &&
-                !line.contains("translated subtitles") &&
-                !line.ends_with("subtitle translation.") &&
-                !line.starts_with("Please note")
-            })
-            .collect();
-            
-        if !filtered_lines.is_empty() {
-            return Some(filtered_lines.join("\n"));
-        }
-    }
-    
-    None
-}
-
-/// Extract translated text from a response string
-/// 
-/// @param response: The response text from the translation service
-/// @param log_capture: Optional log capture for storing warnings instead of printing them
-/// @returns: The extracted translated text or original text as fallback
-fn extract_translated_text_from_response(
-    response: &str, 
-    log_capture: Option<Arc<StdMutex<Vec<LogEntry>>>>
-) -> Result<String> {
-    // If the response is empty, return an error immediately
-    if response.trim().is_empty() {
-        return Err(anyhow!("Empty response received from translation service"));
-    }
-    
-    // Quick check for pure text (most common case in newer implementations)
-    // This avoids unnecessary parsing when the response is already clean
-    if !response.contains('{') && !response.contains('}') && 
-       !response.contains("===") && !response.contains("```") {
-        return Ok(response.trim().to_string());
-    }
-    
-    // Look for the translation block between standard delimiters
-    if let Some(start) = response.find("=== BEGIN TRANSLATION ===") {
-        if let Some(end) = response.find("=== END TRANSLATION ===") {
-            if start < end {
-                let content = &response[start + "=== BEGIN TRANSLATION ===".len()..end];
-                return Ok(content.trim().to_string());
-            }
-        }
-    }
-    
-    // Try to find markdown code blocks (common in newer LLM responses)
-    let code_block_regex = Regex::new(r"```(?:json|text)?\s*\n([\s\S]*?)\n\s*```").unwrap_or_else(|_| {
-        // Fallback to a simpler pattern if the main one fails to compile
-        Regex::new(r"```([\s\S]*?)```").unwrap()
-    });
-    
-    if let Some(caps) = code_block_regex.captures(response) {
-        if let Some(content) = caps.get(1) {
-            return Ok(content.as_str().trim().to_string());
-        }
-    }
-    
-    // Try to extract JSON (common in structured API responses)
-    let json_content = if let (Some(start), Some(end)) = (response.find('{'), response.rfind('}')) {
-        if start < end {
-            Some(&response[start..=end])
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-    
-    if let Some(content) = json_content {
-        // Try to parse as JSON and extract relevant fields
-        if let Ok(json) = serde_json::from_str::<serde_json::Value>(content) {
-            // Check common fields where translation content might be stored
-            for field in ["text", "content", "translation", "translated_text", "result"] {
-                if let Some(text) = json.get(field).and_then(|v| v.as_str()) {
-                    return Ok(text.to_string());
-                }
-            }
-            
-            // Check for array of translations (common pattern in batch responses)
-            for field in ["entries", "translations", "results", "items"] {
-                if let Some(entries) = json.get(field).and_then(|v| v.as_array()) {
-                    let texts: Vec<String> = entries.iter()
-                        .filter_map(|entry| {
-                            // Try common text field names for each entry
-                            for text_field in ["text", "content", "translation", "value"] {
-                                if let Some(text) = entry.get(text_field).and_then(|t| t.as_str()) {
-                                    return Some(text.to_string());
-                                }
-                            }
-                            None
-                        })
-                        .collect();
-                    
-                    if !texts.is_empty() {
-                        return Ok(texts.join("\n"));
-                    }
-                }
-            }
-        }
-    }
-    
-    // Try OpenAI-specific extraction patterns
-    if let Some(extracted) = extract_openai_response(response) {
-        return Ok(extracted);
-    }
-    
-    // If all else fails, just return the original content with warning in logs
-    let warning_message = "Warning: Could not extract structured translation from response, returning raw text. This may affect subtitle alignment.";
-    
-    // Either capture the log or print it directly
-    if let Some(log_capture) = log_capture {
-        let mut logs = match log_capture.lock() {
-            Ok(guard) => guard,
-            Err(poisoned) => {
-                // Get the inner data even though the mutex is poisoned
-                warn!("Encountered poisoned mutex for log capture. Recovering...");
-                poisoned.into_inner()
-            }
-        };
-        logs.push(LogEntry {
-            level: "ERROR".to_string(),
-            message: warning_message.to_string(),
-        });
-    } else {
-        error!("{}", warning_message);
-    }
-    
-    // Return the full original response as a last resort
-    Ok(response.trim().to_string())
-}
-
-/// Log entry for capturing logs during translation
-#[derive(Clone, Debug)]
-pub struct LogEntry {
-    pub level: String,
-    pub message: String,
-}
-
-impl TranslationService {
-    /// Create a new translation service from configuration
-    pub fn new(config: TranslationConfig) -> Result<Self> {
-        let provider = match config.provider {
-            ConfigTranslationProvider::Ollama => {
-                // Parse the Ollama endpoint URL
-                let (host, port) = parse_endpoint(&config.get_endpoint())?;
-                
-                // Create Ollama client
-                let client = Ollama::new(host, port);
-                
-                // Return Ollama provider
-                TranslationProviderImpl::Ollama {
-                    client,
-                }
-            },
-            ConfigTranslationProvider::OpenAI => {
-                // Create OpenAI client
-                let client = OpenAI::new(
-                    config.get_api_key(), 
-                    config.get_endpoint()
-                );
-                
-                // Return OpenAI provider
-                TranslationProviderImpl::OpenAI {
-                    client,
-                }
-            },
-            ConfigTranslationProvider::Anthropic => {
-                // Create Anthropic client
-                let client = Anthropic::new(
-                    config.get_api_key(),
-                    config.get_endpoint()
-                );
-                
-                // Return Anthropic provider
-                TranslationProviderImpl::Anthropic {
-                    client,
-                }
-            },
-        };
-        
-        // Return the initialized service
-        Ok(Self {
-            provider,
-            config,
-        })
-    }
-    
-    /// Test the connection to the translation service
-    pub async fn test_connection(
-        &self, 
-        source_language: &str, 
-        target_language: &str,
-        log_capture: Option<Arc<StdMutex<Vec<LogEntry>>>>
-    ) -> Result<()> {
-        // Test the connection based on the provider
-        match &self.provider {
-            TranslationProviderImpl::Ollama { client } => {
-                let model = self.config.get_model();
-                let prompt = format!("Say hello in {} and {}", source_language, target_language);
-                
-                // Create a simple generation request
-                let request = GenerationRequest::new(model.clone(), prompt)
-                    .temperature(0.7f32);
-                
-                match client.generate(request).await {
-                    Ok(_) => {
-                        Ok(())
-                    },
-                    Err(e) => {
-                        // Capture log or print directly
-                        if let Some(log_capture) = &log_capture {
-                            let mut logs = match log_capture.lock() {
-                                Ok(guard) => guard,
-                                Err(poisoned) => {
-                                    // Get the inner data even though the mutex is poisoned
-                                    warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                    poisoned.into_inner()
-                                }
-                            };
-                            logs.push(LogEntry {
-                                level: "ERROR".to_string(),
-                                message: format!("Ollama connection failed: {}", e)
-                            });
-                        } else {
-                            error!("Ollama connection failed: {}", e);
-                        }
-                        Err(anyhow::anyhow!("Failed to connect to Ollama API: {}", e))
-                    }
-                }
-            },
-            TranslationProviderImpl::OpenAI { client } => {
-                let model = self.config.get_model();
-                
-                // Test connection with a simple prompt
-                match client.test_connection(&model).await {
-                    Ok(_) => {
-                        Ok(())
-                    },
-                    Err(e) => {
-                        // Capture log or print directly
-                        if let Some(log_capture) = &log_capture {
-                            let mut logs = match log_capture.lock() {
-                                Ok(guard) => guard,
-                                Err(poisoned) => {
-                                    // Get the inner data even though the mutex is poisoned
-                                    warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                    poisoned.into_inner()
-                                }
-                            };
-                            logs.push(LogEntry {
-                                level: "ERROR".to_string(),
-                                message: format!("OpenAI connection failed: {}", e)
-                            });
-                        } else {
-                            error!("OpenAI connection failed: {}", e);
-                        }
-                        Err(anyhow::anyhow!("Failed to connect to OpenAI API: {}", e))
-                    }
-                }
-            },
-            TranslationProviderImpl::Anthropic { client } => {
-                let model = self.config.get_model();
-                
-                // Test connection with a simple prompt
-                match client.test_connection(&model).await {
-                    Ok(_) => {
-                        Ok(())
-                    },
-                    Err(e) => {
-                        // Capture log or print directly
-                        if let Some(log_capture) = &log_capture {
-                            let mut logs = match log_capture.lock() {
-                                Ok(guard) => guard,
-                                Err(poisoned) => {
-                                    // Get the inner data even though the mutex is poisoned
-                                    warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                    poisoned.into_inner()
-                                }
-                            };
-                            logs.push(LogEntry {
-                                level: "ERROR".to_string(),
-                                message: format!("Anthropic connection failed: {}", e)
-                            });
-                        } else {
-                            error!("Anthropic connection failed: {}", e);
-                        }
-                        Err(anyhow::anyhow!("Failed to connect to Anthropic API: {}", e))
-                    }
-                }
-            }
-        }
-    }
-    
-    /// Test translation with a sample text
-    pub async fn test_translation(&self, source_language: &str, target_language: &str) -> Result<String> {        
-        let text = format!("Hello, this is a test message. Please translate this from {} to {}.", 
-                          source_language, target_language);
-        self.translate_text(&text, source_language, target_language).await
-    }
-    
-    /// Translate a text from source language to target language
-    async fn translate_text(&self, text: &str, source_language: &str, target_language: &str) -> Result<String> {
-        // Use the with_usage variant and discard the usage stats
-        let (translated_text, _) = self.translate_text_with_usage(text, source_language, target_language, None).await?;
-        Ok(translated_text)
-    }
-    
-    /// Translate a batch of subtitle entries
-    pub async fn translate_batches(&self, 
-                              batches: &[Vec<SubtitleEntry>],
-                              source_language: &str, 
-                              target_language: &str,
-                              log_capture: Arc<StdMutex<Vec<LogEntry>>>,
-                              progress_callback: impl Fn(usize, usize) + Clone + Send + 'static) 
-                              -> Result<(Vec<SubtitleEntry>, TokenUsageStats)> {
-        if batches.is_empty() {
-            // Return early if there are no batches to process
-            return Ok((Vec::new(), TokenUsageStats::with_provider_info(
-                self.config.provider.display_name().to_string(),
-                self.config.get_model().to_string()
-            )));
-        }
-        
-        // Create a collection to store all the final translated entries
-        let all_translated_entries = Arc::new(StdMutex::new(Vec::with_capacity(
-            batches.iter().map(|batch| batch.len()).sum()
-        )));
-        
-        // Create a vector to store the translation tasks
-        let mut tasks = Vec::new();
-        
-        // Get the maximum number of concurrent requests based on configuration
-        let max_concurrent = self.config.optimal_concurrent_requests();
-        
-        // Keep track of the total number of batches
-        let total_batches = batches.len();
-        
-        // Create a counter for completed batches, shared between all tasks
-        let completed = Arc::new(AtomicUsize::new(0));
-        
-        // Create a counter to detect finished processing
-        let processed_batches = Arc::new(AtomicUsize::new(0));
-        
-        // Create a token usage stats tracker shared between all tasks
-        // Initialize with provider and model information
-        let provider_name = self.config.provider.display_name().to_string();
-        let model_name = self.config.get_model().to_string();
-        let token_usage = Arc::new(StdMutex::new(TokenUsageStats::with_provider_info(
-            provider_name,
-            model_name
-        )));
-        
-        let semaphore = Arc::new(Semaphore::new(max_concurrent));
-        
-        // Process each batch
-        for (batch_idx, batch) in batches.iter().enumerate() {
-            if batch.is_empty() {
-                continue;
-            }
-            
-            // Clone references for the task
-            let batch = batch.clone();
-            let source_language = source_language.to_string();
-            let target_language = target_language.to_string();
-            let completed_clone = Arc::clone(&completed);
-            let processed_clone = Arc::clone(&processed_batches);
-            let all_translated_entries_clone = Arc::clone(&all_translated_entries);
-            let progress_callback = progress_callback.clone();
-            let semaphore_clone = Arc::clone(&semaphore);
-            let log_capture_clone = Arc::clone(&log_capture);
-            let translation_service = self.clone();
-            
-            // Create an async task for this batch
-            let task = tokio::spawn(async move {
-                // Acquire a permit from the semaphore to limit concurrent requests
-                let _permit = semaphore_clone.acquire().await.expect("Semaphore should not be closed");
-                
-                // Sleep for rate limit delay to avoid overwhelming the API
-                if batch_idx > 0 {
-                    let delay_ms = translation_service.config.common.rate_limit_delay_ms;
-                    if delay_ms > 0 {
-                        tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                    }
-                }
-                
-                // Translate the batch using the enhanced recovery method
-                let translated_entries = match translation_service.translate_batch_with_recovery(
-                    &batch,
-                    &source_language,
-                    &target_language,
-                    Arc::clone(&log_capture_clone),
-                    true // Enable individual retries for maximum completeness
-                ).await {
-                    Ok(entries) => entries,
-                    Err(e) => {
-                        // Log the error and use originals with warning markers as fallback
-                        let mut logs = log_capture_clone.lock().unwrap();
-                        logs.push(LogEntry {
-                            level: "ERROR".to_string(),
-                            message: format!("Fatal error in batch {}: {}. Using original text with warning markers.", batch_idx + 1, e)
-                        });
-                        
-                        // Mark all entries as needing translation but keep original text
-                        batch.iter().map(|entry| {
-                            let mut new_entry = entry.clone();
-                            new_entry.text = format!("[NEEDS TRANSLATION] {}", entry.text);
-                            new_entry
-                        }).collect()
-                    }
-                };
-                
-                // Update the collection of all translated entries
-                {
-                    let mut all_entries = all_translated_entries_clone.lock().unwrap();
-                    let before_count = all_entries.len();
-                    all_entries.extend(translated_entries);
-                    let after_count = all_entries.len();
-                    
-                    // Log the entry counts for this batch
-                    let mut logs = log_capture_clone.lock().unwrap();
-                    logs.push(LogEntry {
-                        level: "INFO".to_string(),
-                        message: format!("Batch {}: Added {} entries ({} -> {})", 
-                                       batch_idx + 1, after_count - before_count, before_count, after_count)
-                    });
-                }
-                
-                // Update completion tracking
-                let current_completed = completed_clone.fetch_add(1, Ordering::SeqCst) + 1;
-                processed_clone.fetch_add(1, Ordering::SeqCst);
-                
-                // Call progress callback
-                progress_callback(current_completed, total_batches);
-            });
-            
-            tasks.push(task);
-        }
-        
-        // Wait for all translation tasks to complete
-        for task in tasks {
-            let _ = task.await;
-        }
-        
-        // Get the completed entries
-        let mut all_entries = {
-            let entries = all_translated_entries.lock().unwrap();
-            entries.clone()
-        };
-        
-        // Sort the entries by sequence number to maintain proper order
-        all_entries.sort_by_key(|entry| entry.seq_num);
-        
-        // Get the final token usage stats
-        let usage_stats = {
-            let stats = token_usage.lock().unwrap();
-            stats.clone()
-        };
-        
-        // Log final statistics
-        {
-            let mut logs = log_capture.lock().unwrap();
-            let processed = processed_batches.load(Ordering::SeqCst);
-            let expected_entries = batches.iter().map(|batch| batch.len()).sum::<usize>();
-            
-            if all_entries.len() != expected_entries {
-                logs.push(LogEntry {
-                    level: "ERROR".to_string(),
-                    message: format!("MISSING ENTRIES: Expected {} entries but got {}. Some entries may have been lost during translation.", 
-                                  expected_entries, all_entries.len())
-                });
-                
-                // Try to identify which entries are missing
-                let mut expected_seq_nums: Vec<usize> = Vec::new();
-                for batch in batches {
-                    for entry in batch {
-                        expected_seq_nums.push(entry.seq_num);
-                    }
-                }
-                expected_seq_nums.sort();
-                
-                let mut actual_seq_nums: Vec<usize> = all_entries.iter().map(|e| e.seq_num).collect();
-                actual_seq_nums.sort();
-                
-                // Find the missing sequence numbers
-                let mut missing_seq_nums = Vec::new();
-                for seq_num in &expected_seq_nums {
-                    if !actual_seq_nums.contains(seq_num) {
-                        missing_seq_nums.push(*seq_num);
-                    }
-                }
-                
-                if !missing_seq_nums.is_empty() {
-                    logs.push(LogEntry {
-                        level: "ERROR".to_string(),
-                        message: format!("Missing entries with sequence numbers: {:?}", missing_seq_nums)
-                    });
-                }
-            } else {
-                logs.push(LogEntry {
-                    level: "INFO".to_string(),
-                    message: format!("Translation complete: processed {} of {} batches, {} entries", 
-                                  processed, total_batches, all_entries.len())
-                });
-            }
-        }
-        
-        Ok((all_entries, usage_stats))
-    }
-    
-    /// Translate text with usage tracking
-    /// @returns: Translated text and token usage stats (if available)
-    async fn translate_text_with_usage(
-        &self, 
-        text: &str, 
-        source_language: &str, 
-        target_language: &str,
-        log_capture: Option<Arc<StdMutex<Vec<LogEntry>>>>
-    ) -> Result<(String, Option<(Option<u64>, Option<u64>, Option<Duration>)>)> {
-        match &self.provider {
-            TranslationProviderImpl::Ollama { client } => {
-                // Count the number of lines in the text
-                let line_count = text.lines().filter(|line| !line.trim().is_empty()).count();
-                
-                // Enhanced system prompt specifically for subtitles
-                let system_prompt = format!(
-                    "{} You are translating {} subtitles to {}. Each subtitle entry is prefixed with ENTRY_N: where N is the entry number. You MUST preserve these markers and translate EACH entry separately, keeping the exact same format. NEVER merge content between entries. Return each translated entry with its original ENTRY_N: prefix. It is CRITICAL that you process ALL entry numbers and don't skip any.",
-                    self.config.common.system_prompt
-                        .replace("{source_language}", source_language)
-                        .replace("{target_language}", target_language),
-                    source_language,
-                    target_language
-                );
-                
-                // Enhanced user prompt that clearly formats the request
-                let user_prompt = format!(
-                    "Translate the following {} subtitle entries to {}. Each entry is prefixed with ENTRY_N: where N is the entry number.\n\nIMPORTANT RULES:\n- Keep each entry separate with its ENTRY_N: prefix\n- NEVER merge content between entries\n- Translate each entry individually\n- Preserve all formatting\n- Process ALL entries (total: {} entries) - do not skip any\n\nHere are the subtitle entries to translate:\n\n{}",
-                    source_language,
-                    target_language,
-                    text.lines().filter(|line| line.starts_with("ENTRY_")).count(),
-                    text
-                );
-                
-                // Capture log about the request
-                if let Some(log_capture) = &log_capture {
-                    // Handle poisoned mutex gracefully
-                    let mut logs = match log_capture.lock() {
-                        Ok(guard) => guard,
-                        Err(poisoned) => {
-                            // Get the inner data even though the mutex is poisoned
-                            warn!("Encountered poisoned mutex for log capture. Recovering...");
-                            poisoned.into_inner()
-                        }
-                    };
-                    logs.push(LogEntry {
-                        level: "INFO".to_string(),
-                        message: format!("Ollama request: {} lines to translate, {} chars", line_count, text.len())
-                    });
-                }
-                
-                // Build the Ollama request
-                let request = GenerationRequest::new(self.config.get_model(), user_prompt)
-                    .system(system_prompt)
-                    .temperature(self.config.get_temperature())
-                    .no_stream();
-                
-                // Send request to Ollama
-                let request_start = std::time::Instant::now();
-                let response = client.generate(request).await?;
-                let request_duration = request_start.elapsed();
-                
-                // Extract token usage from the response
-                // For Ollama: prompt_eval_count -> prompt_tokens, eval_count -> completion_tokens
-                let prompt_tokens = response.prompt_eval_count;
-                let completion_tokens = response.eval_count;
-                
-                // Extract the response text
-                let response_text = response.response;
-                
-                // Log information about the response
-                if let Some(log_capture) = &log_capture {
-                    let mut logs = match log_capture.lock() {
-                        Ok(guard) => guard,
-                        Err(poisoned) => {
-                            // Get the inner data even though the mutex is poisoned
-                            warn!("Encountered poisoned mutex for log capture. Recovering...");
-                            poisoned.into_inner()
-                        }
-                    };
-                    
-                    // Count lines in the response
-                    let response_line_count = response_text.lines().filter(|line| !line.trim().is_empty()).count();
-                    
-                    logs.push(LogEntry {
-                        level: "INFO".to_string(),
-                        message: format!("Ollama response: got {} lines, expected {} lines", response_line_count, line_count)
-                    });
-                    
-                    // Log first 100 chars of response for debugging
-                    let preview = if response_text.chars().count() > 100 {
-                        // Safely truncate at character boundaries using char iterators
-                        let truncated: String = response_text.chars().take(100).collect();
-                        format!("{}...", truncated)
-                    } else {
-                        response_text.clone()
-                    };
-                    
-                    logs.push(LogEntry {
-                        level: "DEBUG".to_string(),
-                        message: format!("Ollama response preview: {}", preview)
-                    });
-                }
-                
-                // Process the response
-                // Check if we have the right number of lines and adjust if needed
-                let response_lines: Vec<&str> = response_text.lines().filter(|line| !line.trim().is_empty()).collect();
-                let translated_text = if response_lines.len() == line_count {
-                    // Perfect match - use as is
-                    response_text
-                } else if response_lines.len() < line_count {
-                    // Too few lines - log warning and pad with original text
-                    if let Some(log_capture) = &log_capture {
-                        let mut logs = match log_capture.lock() {
-                            Ok(guard) => guard,
-                            Err(poisoned) => {
-                                // Get the inner data even though the mutex is poisoned
-                                warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                poisoned.into_inner()
-                            }
-                        };
-                        logs.push(LogEntry {
-                            level: "WARN".to_string(),
-                            message: format!("Ollama returned fewer lines than expected: got {}, needed {}", 
-                                          response_lines.len(), line_count)
-                        });
-                    }
-                    
-                    // Get original lines
-                    let original_lines: Vec<&str> = text.lines().filter(|line| !line.trim().is_empty()).collect();
-                    
-                    // Create new vector to store final result with missing translations marked
-                    let mut final_lines = Vec::with_capacity(line_count);
-                    
-                    // Add all translated lines (clone to avoid move issue)
-                    for &line in &response_lines {
-                        final_lines.push(line.to_string());
-                    }
-                    
-                    // Add missing lines with marker
-                    for i in response_lines.len()..line_count {
-                        if i < original_lines.len() {
-                            final_lines.push(format!("[MISSING TRANSLATION] {}", original_lines[i]));
-                        } else {
-                            final_lines.push("[MISSING TRANSLATION]".to_string());
-                        }
-                    }
-                    
-                    final_lines.join("\n")
-                } else {
-                    // Too many lines - log warning and use first N lines
-                    if let Some(log_capture) = &log_capture {
-                        let mut logs = match log_capture.lock() {
-                            Ok(guard) => guard,
-                            Err(poisoned) => {
-                                // Get the inner data even though the mutex is poisoned
-                                warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                poisoned.into_inner()
-                            }
-                        };
-                        logs.push(LogEntry {
-                            level: "WARN".to_string(),
-                            message: format!("Ollama returned more lines than expected: got {}, needed {}. Using first {} lines.",
-                                          response_lines.len(), line_count, line_count)
-                        });
-                    }
-                    
-                    response_lines.iter().take(line_count).map(|&s| s.to_string()).collect::<Vec<String>>().join("\n")
-                };
-                
-                // Return the processed translated text
-                Ok((translated_text, Some((prompt_tokens, completion_tokens, Some(request_duration)))))
-            },
-            TranslationProviderImpl::OpenAI { client } => {
-                // Create system prompt
-                let system_prompt = self.config.common.system_prompt
-                    .replace("{source_language}", source_language)
-                    .replace("{target_language}", target_language);
-                
-                // Build the OpenAI request
-                let request = OpenAIRequest::new(self.config.get_model())
-                    .add_message("system", format!("{} Each subtitle entry is prefixed with ENTRY_N: where N is the entry number. You MUST preserve these markers and translate EACH entry separately, keeping the exact same format.", system_prompt))
-                    .add_message("user", format!("Translate the following subtitle entries from {} to {}. Each entry is prefixed with ENTRY_N: where N is the entry number.\n\nIMPORTANT RULES:\n- Keep each entry separate with its ENTRY_N: prefix\n- NEVER merge content between entries\n- Translate each entry individually\n- Preserve all formatting\n- Process ALL entries (total: {} entries) - do not skip any\n\nHere are the subtitle entries to translate:\n\n{}", source_language, target_language, text.lines().filter(|line| line.starts_with("ENTRY_")).count(), text))
-                    .temperature(self.config.get_temperature())
-                    .max_tokens(4096);
-                
-                // Send request to OpenAI
-                let request_start = std::time::Instant::now();
-                let response = client.complete(request).await?;
-                let request_duration = request_start.elapsed();
-                
-                // Extract the content from the first choice
-                if response.choices.is_empty() {
-                    return Err(anyhow!("No response received from OpenAI API"));
-                }
-                
-                // Extract token usage information 
-                // For OpenAI: prompt_tokens -> prompt_tokens, completion_tokens -> completion_tokens
-                let prompt_tokens = Some(response.usage.prompt_tokens as u64);
-                let completion_tokens = Some(response.usage.completion_tokens as u64);
-                
-                // Extract the translated text from the response
-                Ok((extract_translated_text_from_response(&response.choices[0].message.content, log_capture)?, Some((prompt_tokens, completion_tokens, Some(request_duration)))))
-            },
-            TranslationProviderImpl::Anthropic { client } => {
-                // Calculate appropriate max tokens (approximately 6 times the input length to allow for expansion)
-                let text_length = text.len();
-                
-                // Create system prompt
-                let system_prompt = self.config.common.system_prompt
-                    .replace("{source_language}", source_language)
-                    .replace("{target_language}", target_language);
-                
-                // Determine the maximum allowed tokens based on the model
-                let model_max_tokens = self.max_tokens_for_model(&self.config.get_model());
-                
-                // Use a more generous multiplier for Anthropic to ensure we get complete translations
-                // but ensure we respect the model's token limit
-                let max_tokens = ((text_length * 6).max(1000).min(model_max_tokens as usize)) as u32;
-                
-                // Count the exact number of non-empty subtitle lines
-                let subtitle_lines = text.split('\n').filter(|s| !s.is_empty()).collect::<Vec<_>>();
-                let input_lines_count = subtitle_lines.len();
-                
-                // Capture log instead of directly printing
-                if let Some(log_capture) = &log_capture {
-                    let mut logs = match log_capture.lock() {
-                        Ok(guard) => guard,
-                        Err(poisoned) => {
-                            // Get the inner data even though the mutex is poisoned
-                            warn!("Encountered poisoned mutex for log capture. Recovering...");
-                            poisoned.into_inner()
-                        }
-                    };
-                    logs.push(LogEntry {
-                        level: "DEBUG".to_string(),
-                        message: format!("Anthropic model: {}, Input size: {} chars, {} entries, {} lines, max tokens: {}", 
-                                    self.config.get_model(), text_length, text.lines().filter(|line| line.starts_with("ENTRY_")).count(), input_lines_count, max_tokens)
-                    });
-                }
-                
-                // Add specific instructions to enhance the system prompt for subtitle translation
-                let enhanced_system_prompt = format!(
-                    "{} You are a professional subtitle translator specializing in {}-to-{} translation. CRITICAL: You MUST process EVERY ENTRY_N: marker in the input and never skip any entries. Always maintain the ENTRY_N: prefix format for each entry.",
-                    system_prompt,
-                    source_language,
-                    target_language
-                );
-                
-                // Build the Anthropic request with enhanced instructions
-                let request = AnthropicRequest::new(self.config.get_model(), max_tokens)
-                    .system(enhanced_system_prompt)
-                    .add_message("user", format!(
-                        "Translate the following subtitle entries from {} to {}.\n\nIMPORTANT RULES:\n- Each entry is prefixed with ENTRY_N: where N is the entry number\n- Keep each entry separate with its ENTRY_N: prefix\n- NEVER merge content between entries\n- Translate each entry individually\n- Preserve all formatting\n- Process ALL entries (total: {} entries) - do not skip any\n\nHere are the subtitle entries to translate:\n\n{}", 
-                        source_language, 
-                        target_language,
-                        text.lines().filter(|line| line.starts_with("ENTRY_")).count(),
-                        text
-                    ))
-                    .temperature(self.config.get_temperature());
-                
-                // Send request to Anthropic with timing
-                let request_start = std::time::Instant::now();
-                let response = match client.complete(request).await {
-                    Ok(resp) => resp,
-                    Err(e) => {
-                        // Capture error log instead of directly printing
-                        if let Some(log_capture) = &log_capture {
-                            let mut logs = match log_capture.lock() {
-                                Ok(guard) => guard,
-                                Err(poisoned) => {
-                                    // Get the inner data even though the mutex is poisoned
-                                    warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                    poisoned.into_inner()
-                                }
-                            };
-                            logs.push(LogEntry {
-                                level: "ERROR".to_string(),
-                                message: format!("Anthropic request failed after {:?}: {}", request_start.elapsed(), e)
-                            });
-                        } else {
-                            error!("Anthropic request failed after {:?}: {}", request_start.elapsed(), e);
-                        }
-                        return Err(e);
-                    }
-                };
-                let request_duration = request_start.elapsed();
-                
-                // Capture log for response time
-                if let Some(log_capture) = &log_capture {
-                    let mut logs = match log_capture.lock() {
-                        Ok(guard) => guard,
-                        Err(poisoned) => {
-                            // Get the inner data even though the mutex is poisoned
-                            warn!("Encountered poisoned mutex for log capture. Recovering...");
-                            poisoned.into_inner()
-                        }
-                    };
-                    logs.push(LogEntry {
-                        level: "INFO".to_string(),
-                        message: format!("Anthropic response received in {:?}", request_duration)
-                    });
-                } else {
-                    info!("Anthropic response received in {:?}", request_duration);
-                }
-                
-                // Extract token usage information
-                // For Anthropic: input_tokens -> prompt_tokens, output_tokens -> completion_tokens
-                let prompt_tokens = Some(response.usage.input_tokens as u64);
-                let completion_tokens = Some(response.usage.output_tokens as u64);
-                
-                // Extract the text from the response
-                let content = Anthropic::extract_text_from_response(&response);
-                
-                // Process the response to ensure we get the correct number of lines
-                let mut output_lines = content.split('\n')
-                    .map(|s| s.trim().to_string())
-                    .filter(|s| !s.is_empty())
-                    .collect::<Vec<String>>();
-                
-                let output_lines_count = output_lines.len();
-                
-                // Capture log for response line count
-                if let Some(log_capture) = &log_capture {
-                    let mut logs = match log_capture.lock() {
-                        Ok(guard) => guard,
-                        Err(poisoned) => {
-                            // Get the inner data even though the mutex is poisoned
-                            warn!("Encountered poisoned mutex for log capture. Recovering...");
-                            poisoned.into_inner()
-                        }
-                    };
-                    logs.push(LogEntry {
-                        level: "INFO".to_string(),
-                        message: format!("Anthropic translation: got {} lines, expected {} lines", 
-                                        output_lines_count, input_lines_count)
-                    });
-                } else {
-                    info!("Anthropic translation: got {} lines, expected {} lines", output_lines_count, input_lines_count);
-                }
-                
-                let final_content = if output_lines_count == input_lines_count {
-                    // Perfect match - use as is
-                    content
-                } else if output_lines_count < input_lines_count {
-                    // Not enough lines - log warning and pad with missing translation markers
-                    if let Some(log_capture) = &log_capture {
-                        let mut logs = match log_capture.lock() {
-                            Ok(guard) => guard,
-                            Err(poisoned) => {
-                                // Get the inner data even though the mutex is poisoned
-                                warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                poisoned.into_inner()
-                            }
-                        };
-                        logs.push(LogEntry {
-                            level: "ERROR".to_string(),
-                            message: format!("Translation returned fewer lines than input: got {} but expected {}.", 
-                                           output_lines_count, input_lines_count)
-                        });
-                    } else {
-                        error!("Translation returned fewer lines than input: got {} but expected {}.", 
-                               output_lines_count, input_lines_count);
-                    }
-                    
-                    // Extend output_lines vector with missing translation markers for original lines
-                    while output_lines.len() < input_lines_count {
-                        let original_idx = output_lines.len();
-                        if original_idx < subtitle_lines.len() {
-                            // Add missing translation marker with original text
-                            output_lines.push(format!("[MISSING TRANSLATION] {}", subtitle_lines[original_idx]));
-                        } else {
-                            // Failsafe - should never happen but just in case
-                            output_lines.push("[MISSING TRANSLATION]".to_string());
-                        }
-                    }
-                    
-                    // Join the lines back together
-                    output_lines.join("\n")
-                } else {
-                    // Too many lines - log and use the smart mapping algorithm in translate_batches
-                    if let Some(log_capture) = &log_capture {
-                        let mut logs = match log_capture.lock() {
-                            Ok(guard) => guard,
-                            Err(poisoned) => {
-                                // Get the inner data even though the mutex is poisoned
-                                warn!("Encountered poisoned mutex for log capture. Recovering...");
-                                poisoned.into_inner()
-                            }
-                        };
-                        logs.push(LogEntry {
-                            level: "INFO".to_string(),
-                            message: format!("Translation returned more lines than input: got {} but expected {}.", 
-                                           output_lines_count, input_lines_count)
-                        });
-                    } else {
-                        info!("Translation returned more lines than input: got {} but expected {}.", 
-                               output_lines_count, input_lines_count);
-                    }
-                    
-                    // Keep the content as is, the batch processing will handle mapping extra lines
-                    content
-                };
-                
-                // Extract the translated text from the response
-                Ok((extract_translated_text_from_response(&final_content, log_capture)?, Some((prompt_tokens, completion_tokens, Some(request_duration)))))
-            }
-        }
-    }
-
-    /// Process a batch of subtitles with enhanced error recovery
-    pub async fn translate_batch_with_recovery(
-        &self,
-        batch: &[SubtitleEntry],
-        source_language: &str,
-        target_language: &str,
-        log_capture: Arc<StdMutex<Vec<LogEntry>>>,
-        retry_individual_entries: bool
-    ) -> Result<Vec<SubtitleEntry>> {
-        // Capture log information
-        let capture_log = |level: &str, message: &str| {
-            let mut logs = log_capture.lock().unwrap();
-            logs.push(LogEntry {
-                level: level.to_string(),
-                message: message.to_string()
-            });
-        };
-        
-        // Track entries with special formatting that might need special handling
-        let entries_with_special_formatting: Vec<usize> = batch.iter()
-            .filter(|entry| {
-                // Check for format codes or tags that might confuse the LLM
-                // SRT format supports these HTML-like tags: <b>, <i>, <u>, <font color="...">, and positioning codes like {\an8}
-                entry.text.contains("{\\an") ||   // Position tags like {\an8}
-                entry.text.contains("<i>") || entry.text.contains("</i>") ||  // Italic
-                entry.text.contains("<b>") || entry.text.contains("</b>") ||  // Bold
-                entry.text.contains("<u>") || entry.text.contains("</u>") ||  // Underline
-                entry.text.contains("<font") || entry.text.contains("</font>") ||  // Font color
-                entry.text.contains("{i}") || entry.text.contains("{/i}") ||  // Alternative italic syntax
-                entry.text.contains("{b}") || entry.text.contains("{/b}") ||  // Alternative bold syntax
-                entry.text.contains("{u}") || entry.text.contains("{/u}") ||  // Alternative underline syntax
-                entry.text.contains("♪") ||  // Music notes, common in subtitles
-                // Check for unbalanced formatting tags or bracket usage
-                entry.text.matches('<').count() != entry.text.matches('>').count() ||
-                entry.text.matches('{').count() != entry.text.matches('}').count()
-            })
-            .map(|entry| entry.seq_num)
-            .collect();
-        
-        // Log entries with special formatting for diagnostic purposes
-        if !entries_with_special_formatting.is_empty() {
-            capture_log("INFO", &format!("Batch contains {} entries with special formatting that may require individual handling", 
-                entries_with_special_formatting.len()));
-        }
-        
-        // Prepare text for translation with entry markers
-        let prepare_text = |entries: &[SubtitleEntry]| -> String {
-            entries.iter()
-                .map(|entry| {
-                    if entries_with_special_formatting.contains(&entry.seq_num) {
-                        // For entries with special formatting, add extra marker to make it more clear
-                        format!("ENTRY_{}:\n# SPECIAL_FORMATTING #\n{}", entry.seq_num, entry.text)
-                    } else {
-                        format!("ENTRY_{}:\n{}", entry.seq_num, entry.text)
-                    }
-                })
-                .collect::<Vec<_>>()
-                .join("\n\n")
-        };
-        
-        // First attempt: batch translation
-        let text_to_translate = prepare_text(batch);
-        
-        // Try batch translation with up to 3 retries
-        let mut batch_success = false;
-        let mut translated_text = String::new();
-        let mut retry_count = 0;
-        let max_retries = 3;
-        
-        while !batch_success && retry_count < max_retries {
-            match self.translate_text_with_usage(
-                &text_to_translate,
-                source_language,
-                target_language,
-                Some(Arc::clone(&log_capture))
-            ).await {
-                Ok((result, _)) => {
-                    translated_text = result;
-                    batch_success = true;
-                    break;
-                },
-                Err(e) => {
-                    retry_count += 1;
-                    capture_log("WARN", &format!("Batch translation failed (attempt {}/{}): {}", 
-                              retry_count, max_retries, e));
-                    
-                    // Wait before retrying with increasing delay
-                    if retry_count < max_retries {
-                        let delay_ms = 1000 * retry_count;
-                        tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                    }
-                }
-            }
-        }
-        
-        // If batch translation succeeded, extract entries
-        if batch_success {
-            // Extract entries from the translated text based on entry markers
-            let mut entries_map = std::collections::HashMap::new();
-            let mut current_entry_num: Option<usize> = None;
-            let mut current_entry_text = String::new();
-            
-            // Process each line to find entry markers and extract content
-            for line in translated_text.lines() {
-                // Try different entry marker formats that the LLM might return
-                let entry_marker = if let Some(marker) = line.strip_prefix("ENTRY_") {
-                    Some(marker)
-                } else if let Some(marker) = line.strip_prefix("Entry_") {
-                    Some(marker)
-                } else if let Some(marker) = line.strip_prefix("ENTRY ") {
-                    Some(marker)
-                } else if let Some(marker) = line.strip_prefix("Entry ") {
-                    Some(marker)
-                } else {
-                    None
-                };
-                
-                if let Some(entry_marker) = entry_marker {
-                    // If we were already processing an entry, save it
-                    if let Some(entry_num) = current_entry_num {
-                        if !current_entry_text.is_empty() {
-                            entries_map.insert(entry_num, current_entry_text.trim().to_string());
-                            current_entry_text = String::new();
-                        }
-                    }
-                    
-                    // Parse the new entry number with more robust handling
-                    // Try to find any number in the marker and use that as entry number
-                    if let Some(colon_idx) = entry_marker.find(':') {
-                        if let Ok(entry_num) = entry_marker[..colon_idx].trim().parse::<usize>() {
-                            current_entry_num = Some(entry_num);
-                            continue;
-                        }
-                    } else {
-                        // Try to extract just the number from the marker string
-                        let number_str: String = entry_marker.chars()
-                            .take_while(|c| c.is_ascii_digit())
-                            .collect();
-                        
-                        if !number_str.is_empty() {
-                            if let Ok(entry_num) = number_str.parse::<usize>() {
-                                current_entry_num = Some(entry_num);
-                                continue;
-                            }
-                        }
-                    }
-                } else if let Some(_entry_num) = current_entry_num {
-                    // Skip the special formatting marker
-                    if line == "# SPECIAL_FORMATTING #" {
-                        continue;
-                    }
-                    
-                    // Add content to the current entry
-                    if !current_entry_text.is_empty() {
-                        current_entry_text.push('\n');
-                    }
-                    current_entry_text.push_str(line);
-                }
-            }
-            
-            // Add the last entry if needed
-            if let Some(entry_num) = current_entry_num {
-                if !current_entry_text.is_empty() {
-                    entries_map.insert(entry_num, current_entry_text.trim().to_string());
-                }
-            }
-            
-            // Fallback method for entries: try to find entries that weren't properly formatted
-            // This helps catch entries where the LLM might have reformatted the markers
-            if entries_map.len() < batch.len() {
-                // Collect seq_nums that we're looking for but haven't found yet
-                let missing_seq_nums: Vec<usize> = batch.iter()
-                    .map(|entry| entry.seq_num)
-                    .filter(|seq_num| !entries_map.contains_key(seq_num))
-                    .collect();
-                
-                if !missing_seq_nums.is_empty() {
-                    // Try a regex-based approach to find entry markers
-                    for missing_num in &missing_seq_nums {
-                        // Try different patterns that might indicate entry markers
-                        let patterns = [
-                            format!(r"(?i)ENTRY[_\s]*{}[:\s]", missing_num),  // ENTRY_123:, Entry 123:, etc.
-                            format!(r"(?i)ENTRY[_\s]*\[{}]", missing_num),    // ENTRY[123]
-                            format!(r"(?i)ENTRY[_\s]*#{}[:\s]", missing_num), // ENTRY #123:
-                            format!(r"(?i)ENTRY[_\s]*NUMBER[_\s]*{}[:\s]", missing_num), // ENTRY NUMBER 123:
-                        ];
-                        
-                        for pattern in &patterns {
-                            // Look for the pattern in the translated text
-                            if let Ok(regex) = regex::Regex::new(pattern) {
-                                if let Some(mat) = regex.find(&translated_text) {
-                                    // Find where the entry starts and the next entry starts
-                                    let entry_start = mat.end();
-                                    
-                                    // Find the next entry marker or end of text
-                                    let next_entry_start = patterns.iter()
-                                        .filter_map(|p| {
-                                            if let Ok(next_regex) = regex::Regex::new(p) {
-                                                next_regex.find_iter(&translated_text[entry_start..])
-                                                    .next()
-                                                    .map(|m| entry_start + m.start())
-                                            } else {
-                                                None
-                                            }
-                                        })
-                                        .min()
-                                        .unwrap_or(translated_text.len());
-                                    
-                                    // Extract the entry text
-                                    let entry_text = translated_text[entry_start..next_entry_start].trim();
-                                    if !entry_text.is_empty() {
-                                        entries_map.insert(*missing_num, entry_text.to_string());
-                                        break; // Found this entry, move to the next
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            
-            // Create subtitle entries from the map
-            let mut result_entries = Vec::with_capacity(batch.len());
-            
-            // Check if we have all entries or need individual translation
-            let all_entries_present = batch.iter()
-                .all(|entry| entries_map.contains_key(&entry.seq_num));
-            
-            if all_entries_present {
-                // All entries were translated successfully
-                for entry in batch {
-                    let mut translated_entry = entry.clone();
-                    let translated_text = entries_map.get(&entry.seq_num).unwrap();
-                    translated_entry.text = preserve_formatting(&entry.text, translated_text);
-                    result_entries.push(translated_entry);
-                }
-                
-                capture_log("INFO", &format!("Successfully translated all {} entries in batch", batch.len()));
-                return Ok(result_entries);
-            }
-            
-            if retry_individual_entries {
-                // Some entries are missing - translate them individually
-                // First, try entries with special formatting that might have been missed
-                for entry in batch {
-                    if entries_with_special_formatting.contains(&entry.seq_num) && !entries_map.contains_key(&entry.seq_num) {
-                        // Proactively translate specially formatted entries
-                        capture_log("INFO", &format!("Proactively translating entry {} with special formatting", entry.seq_num));
-                        
-                        // Individual translation with retries
-                        let individual_text = format!("ENTRY_{}:\n{}", entry.seq_num, entry.text);
-                        let mut individual_success = false;
-                        let mut individual_result = String::new();
-                        let mut retry_count = 0;
-                        
-                        // Use smaller token limit for individual translations to avoid issues
-                        let mut reduce_tokens_factor = 0.8;
-                        
-                        while !individual_success && retry_count < max_retries {
-                            match self.translate_text_with_usage(
-                                &individual_text,
-                                source_language,
-                                target_language,
-                                Some(Arc::clone(&log_capture))
-                            ).await {
-                                Ok((result, _)) => {
-                                    individual_result = result;
-                                    individual_success = true;
-                                    break;
-                                },
-                                Err(e) => {
-                                    retry_count += 1;
-                                    // If the error might be related to token limits, reduce the token count
-                                    if e.to_string().contains("token") {
-                                        reduce_tokens_factor *= 0.8;
-                                    }
-                                    
-                                    capture_log("WARN", &format!("Individual translation for entry {} failed (attempt {}/{}): {}", 
-                                                entry.seq_num, retry_count, max_retries, e));
-                                    
-                                    if retry_count < max_retries {
-                                        let delay_ms = 500 * retry_count;
-                                        tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                                    }
-                                }
-                            }
-                        }
-                        
-                        if individual_success {
-                            // Extract the content from the individual translation
-                            let content = individual_result.lines()
-                                .skip_while(|line| !line.starts_with("ENTRY_"))
-                                .skip(1) // Skip the marker line
-                                .filter(|line| *line != "# SPECIAL_FORMATTING #") // Filter out the special formatting marker
-                                .collect::<Vec<&str>>()
-                                .join("\n");
-                                
-                            if !content.trim().is_empty() {
-                                // Store the result in our entries map for later use
-                                entries_map.insert(entry.seq_num, content.trim().to_string());
-                                capture_log("INFO", &format!("Successfully pre-translated entry {} with special formatting", entry.seq_num));
-                            }
-                        }
-                    }
-                }
-                
-                // Now process all entries, including the pre-translated special ones
-                for entry in batch {
-                    let mut translated_entry = entry.clone();
-                    
-                    if let Some(translated_text) = entries_map.get(&entry.seq_num) {
-                        // Entry was translated successfully in the batch
-                        translated_entry.text = preserve_formatting(&entry.text, translated_text);
-                    } else {
-                        // Entry missing - try individual translation
-                        capture_log("WARN", &format!("Entry {} missing from batch translation. Attempting individual translation...", entry.seq_num));
-                        
-                        // Individual translation with retries
-                        let individual_text = format!("ENTRY_{}:\n{}", entry.seq_num, entry.text);
-                        let mut individual_success = false;
-                        let mut individual_result = String::new();
-                        let mut retry_count = 0;
-                        
-                        // Start with normal max_tokens calculation, but reduce if we get token limit errors
-                        let mut reduce_tokens_factor = 1.0;
-                        
-                        while !individual_success && retry_count < max_retries {
-                            // Calculate max_tokens with potential reduction factor
-                            let _text_len = individual_text.len();
-                            let mut _is_token_limit_error = false;
-                            
-                            match self.translate_text_with_usage(
-                                &individual_text,
-                                source_language,
-                                target_language,
-                                Some(Arc::clone(&log_capture))
-                            ).await {
-                                Ok((result, _)) => {
-                                    individual_result = result;
-                                    individual_success = true;
-                                    break;
-                                },
-                                Err(e) => {
-                                    retry_count += 1;
-                                    
-                                    // Check if this is a max_tokens error for Anthropic
-                                    let error_msg = e.to_string().to_lowercase();
-                                    _is_token_limit_error = error_msg.contains("max_tokens") && 
-                                                      (error_msg.contains("maximum allowed") || 
-                                                       error_msg.contains("exceeds"));
-                                    
-                                    if _is_token_limit_error && retry_count < max_retries {
-                                        // This is a token limit error, adjust our max_tokens calculation
-                                        reduce_tokens_factor *= 0.7; // Reduce by 30% each time
-                                        capture_log("WARN", &format!("Token limit error for entry {}. Reducing max_tokens by factor {}.", 
-                                                 entry.seq_num, reduce_tokens_factor));
-                                    } else {
-                                        // Not a token error or we've already retried too many times
-                                        capture_log("WARN", &format!("Individual translation for entry {} failed (attempt {}/{}): {}", 
-                                                 entry.seq_num, retry_count, max_retries, e));
-                                    }
-                                    
-                                    // Wait before retrying
-                                    if retry_count < max_retries {
-                                        let delay_ms = 1000 * retry_count;
-                                        tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                                    }
-                                }
-                            }
-                        }
-                        
-                        if individual_success {
-                            // Extract the content from the individual translation
-                            let content = individual_result.lines()
-                                .skip_while(|line| !line.starts_with("ENTRY_"))
-                                .skip(1) // Skip the marker line
-                                .filter(|line| *line != "# SPECIAL_FORMATTING #") // Filter out the special formatting marker
-                                .collect::<Vec<&str>>()
-                                .join("\n");
-                                
-                            if !content.trim().is_empty() {
-                                translated_entry.text = preserve_formatting(&entry.text, content.trim());
-                                capture_log("INFO", &format!("Successfully translated entry {} individually", entry.seq_num));
-                            } else {
-                                // Couldn't extract proper content - use original with warning
-                                capture_log("ERROR", &format!("Failed to extract translated content for entry {}. Using original text.", entry.seq_num));
-                                translated_entry.text = entry.text.to_string();
-                            }
-                        } else {
-                            // Individual translation failed - use original with warning
-                            capture_log("ERROR", &format!("All individual translation attempts for entry {} failed. Using original with warning.", entry.seq_num));
-                            translated_entry.text = format!("[NEEDS TRANSLATION] {}", entry.text);
-                        }
-                    }
-                    
-                    result_entries.push(translated_entry);
-                }
-                
-                return Ok(result_entries);
-            } else {
-                // Some entries are missing but individual retry is disabled
-                for entry in batch {
-                    let mut translated_entry = entry.clone();
-                    
-                    if let Some(translated_text) = entries_map.get(&entry.seq_num) {
-                        // Entry was translated successfully
-                        translated_entry.text = preserve_formatting(&entry.text, translated_text);
-                    } else {
-                        // Entry missing - use original with warning
-                        capture_log("ERROR", &format!("Entry {} missing from translation and individual retry disabled. Using original with warning.", entry.seq_num));
-                        translated_entry.text = format!("[NEEDS TRANSLATION] {}", entry.text);
-                    }
-                    
-                    result_entries.push(translated_entry);
-                }
-                
-                return Ok(result_entries);
-            }
-        }
-        
-        // Batch translation failed entirely - use individual translation if enabled
-        if retry_individual_entries {
-            // Translate each entry individually
-            let mut result_entries = Vec::with_capacity(batch.len());
-            
-            for entry in batch {
-                let mut translated_entry = entry.clone();
-                let individual_text = format!("ENTRY_{}:\n{}", entry.seq_num, entry.text);
-                
-                // Try individual translation with retries
-                let mut individual_success = false;
-                let mut individual_result = String::new();
-                let mut retry_count = 0;
-                
-                // Start with normal max_tokens calculation, but reduce if we get token limit errors
-                let mut reduce_tokens_factor = 1.0;
-                
-                while !individual_success && retry_count < max_retries {
-                    // Calculate max_tokens with potential reduction factor
-                    let _text_len = individual_text.len();
-                    let mut _is_token_limit_error = false;
-                    
-                    match self.translate_text_with_usage(
-                        &individual_text,
-                        source_language,
-                        target_language,
-                        Some(Arc::clone(&log_capture))
-                    ).await {
-                        Ok((result, _)) => {
-                            individual_result = result;
-                            individual_success = true;
-                            break;
-                        },
-                        Err(e) => {
-                            retry_count += 1;
-                            
-                            // Check if this is a max_tokens error for Anthropic
-                            let error_msg = e.to_string().to_lowercase();
-                            _is_token_limit_error = error_msg.contains("max_tokens") && 
-                                                  (error_msg.contains("maximum allowed") || 
-                                                   error_msg.contains("exceeds"));
-                            
-                            if _is_token_limit_error && retry_count < max_retries {
-                                // This is a token limit error, adjust our max_tokens calculation
-                                reduce_tokens_factor *= 0.7; // Reduce by 30% each time
-                                capture_log("WARN", &format!("Token limit error for entry {}. Reducing max_tokens by factor {}.", 
-                                         entry.seq_num, reduce_tokens_factor));
-                            } else {
-                                // Not a token error or we've already retried too many times
-                                capture_log("WARN", &format!("Individual translation for entry {} failed (attempt {}/{}): {}", 
-                                         entry.seq_num, retry_count, max_retries, e));
-                            }
-                            
-                            // Wait before retrying
-                            if retry_count < max_retries {
-                                let delay_ms = 1000 * retry_count;
-                                tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                            }
-                        }
-                    }
-                }
-                
-                if individual_success {
-                    // Extract the content from the individual translation
-                    let content = individual_result.lines()
-                        .skip_while(|line| !line.starts_with("ENTRY_"))
-                        .skip(1) // Skip the marker line
-                        .filter(|line| *line != "# SPECIAL_FORMATTING #") // Filter out the special formatting marker
-                        .collect::<Vec<&str>>()
-                        .join("\n");
-                        
-                    if !content.trim().is_empty() {
-                        translated_entry.text = preserve_formatting(&entry.text, content.trim());
-                        capture_log("INFO", &format!("Successfully translated entry {} individually", entry.seq_num));
-                    } else {
-                        // Couldn't extract proper content - use original with warning
-                        capture_log("ERROR", &format!("Failed to extract translated content for entry {} after batch failure. Using original text.", entry.seq_num));
-                        translated_entry.text = entry.text.to_string();
-                    }
-                } else {
-                    // Individual translation failed - use original with warning
-                    capture_log("ERROR", &format!("All translation attempts for entry {} failed. Using original with warning.", entry.seq_num));
-                    translated_entry.text = format!("[NEEDS TRANSLATION] {}", entry.text);
-                }
-                
-                result_entries.push(translated_entry);
-            }
-            
-            return Ok(result_entries);
-        } else {
-            // Batch translation failed and individual retry is disabled - use all originals with warnings
-            Ok(batch.iter().map(|entry| {
-                let mut new_entry = entry.clone();
-                capture_log("ERROR", &format!("Batch translation failed and individual retry disabled for entry {}. Using original with warning.", entry.seq_num));
-                new_entry.text = format!("[NEEDS TRANSLATION] {}", entry.text);
-                new_entry
-            }).collect())
-        }
-    }
-
-    /// Get the maximum allowed tokens for a model
-    fn max_tokens_for_model(&self, model: &str) -> u32 {
-        // Extract model name from the config
-        let model_name = model.to_lowercase();
-        
-        // Determine max tokens based on model name
-        if model_name.contains("haiku") {
-            // Claude-3-Haiku has a limit of 4096 tokens
-            4096
-        } else if model_name.contains("sonnet") {
-            // Claude-3-Sonnet has a higher limit
-            12288
-        } else if model_name.contains("opus") {
-            // Claude-3-Opus has the highest limit
-            32768
-        } else if model_name.contains("gpt-4") {
-            // GPT-4 family typically has high token limits
-            8192
-        } else if model_name.contains("gpt-3.5") {
-            // GPT-3.5 models like turbo
-            4096
-        } else if model_name.contains("mixtral") || model_name.contains("llama") {
-            // Mixtral, Llama etc. have flexible limits but we'll be conservative
-            4096
-        } else {
-            // Default fallback for unknown models
-            4096
-        }
-    }
-}
-
-impl Clone for TranslationService {
-    fn clone(&self) -> Self {
-        // We need to manually clone because Ollama client doesn't implement Clone
-        let provider = match &self.provider {
-            TranslationProviderImpl::Ollama { client: _ } => {
-                // Parse the Ollama endpoint URL
-                let (host, port) = parse_endpoint(&self.config.get_endpoint())
-                    .expect("Failed to parse endpoint URL");
-                
-                // Create a new Ollama client
-                let client = Ollama::new(host, port);
-                
-                TranslationProviderImpl::Ollama {
-                    client,
-                }
-            },
-            TranslationProviderImpl::OpenAI { client: _ } => {
-                // Create a new OpenAI client
-                let client = OpenAI::new(
-                    self.config.get_api_key(),
-                    self.config.get_endpoint()
-                );
-                
-                TranslationProviderImpl::OpenAI {
-                    client,
-                }
-            },
-            TranslationProviderImpl::Anthropic { client: _ } => {
-                // Create a new Anthropic client
-                let client = Anthropic::new(
-                    self.config.get_api_key(),
-                    self.config.get_endpoint()
-                );
-                
-                TranslationProviderImpl::Anthropic {
-                    client,
-                }
-            },
-        };
-        
-        Self {
-            provider,
-            config: self.config.clone(),
-        }
-    }
-}
-
-/// Helper function to preserve formatting markers while replacing content
-fn preserve_formatting(original: &str, translated: &str) -> String {
-    // Identify and preserve special formatting codes in the translation
-    
-    // First, extract all special formatting codes from the original
-    // Position codes - {\anN}
-    let an_codes = regex::Regex::new(r"\{\\an\d\}").unwrap();
-    
-    // HTML-style formatting tags
-    let italic_tags = regex::Regex::new(r"<i>|</i>").unwrap();
-    let bold_tags = regex::Regex::new(r"<b>|</b>").unwrap();
-    let underline_tags = regex::Regex::new(r"<u>|</u>").unwrap();
-    let font_tags = regex::Regex::new(r"<font.*?>|</font>").unwrap();
-    
-    // Alternative formatting tags in curly braces
-    let alt_italic_tags = regex::Regex::new(r"\{i\}|\{/i\}").unwrap();
-    let alt_bold_tags = regex::Regex::new(r"\{b\}|\{/b\}").unwrap();
-    let alt_underline_tags = regex::Regex::new(r"\{u\}|\{/u\}").unwrap();
-    
-    // Other format codes in curly braces
-    let format_codes = regex::Regex::new(r"\{[^}]+\}").unwrap();
-    
-    // Music note symbols
-    let music_symbols = regex::Regex::new(r"♪").unwrap();
-    
-    // Extract format codes from original
-    let mut original_codes = Vec::new();
-    
-    // Collect all {\anX} position codes
-    for cap in an_codes.find_iter(original) {
-        original_codes.push((cap.start(), cap.end(), cap.as_str().to_string()));
-    }
-    
-    // Collect all HTML-style formatting tags
-    for tag_regex in [&italic_tags, &bold_tags, &underline_tags, &font_tags] {
-        for cap in tag_regex.find_iter(original) {
-            original_codes.push((cap.start(), cap.end(), cap.as_str().to_string()));
-        }
-    }
-    
-    // Collect all alternative formatting tags in curly braces
-    for tag_regex in [&alt_italic_tags, &alt_bold_tags, &alt_underline_tags] {
-        for cap in tag_regex.find_iter(original) {
-            original_codes.push((cap.start(), cap.end(), cap.as_str().to_string()));
-        }
-    }
-    
-    // Collect all other format codes in curly braces
-    for cap in format_codes.find_iter(original) {
-        // Skip if already captured as an an_code or alt tag
-        if !an_codes.is_match(cap.as_str()) && 
-           !alt_italic_tags.is_match(cap.as_str()) && 
-           !alt_bold_tags.is_match(cap.as_str()) && 
-           !alt_underline_tags.is_match(cap.as_str()) {
-            original_codes.push((cap.start(), cap.end(), cap.as_str().to_string()));
-        }
-    }
-    
-    // Collect music symbols (♪)
-    for cap in music_symbols.find_iter(original) {
-        original_codes.push((cap.start(), cap.end(), cap.as_str().to_string()));
-    }
-    
-    // Sort codes by position in the original text
-    original_codes.sort_by_key(|(start, _, _)| *start);
-    
-    // If there are no format codes, just return the translated text
-    if original_codes.is_empty() {
-        return translated.to_string();
-    }
-    
-    // Check if the LLM already kept the formatting codes intact
-    let mut translated_contains_all_codes = true;
-    for (_, _, code) in &original_codes {
-        if !translated.contains(code) {
-            translated_contains_all_codes = false;
-            break;
-        }
-    }
-    
-    // If all codes are already in the translation, just return it
-    if translated_contains_all_codes {
-        return translated.to_string();
-    }
-    
-    // Map of opening tags to their closing tags
-    let mut tag_pairs = std::collections::HashMap::new();
-    tag_pairs.insert("<i>".to_string(), "</i>".to_string());
-    tag_pairs.insert("<b>".to_string(), "</b>".to_string());
-    tag_pairs.insert("<u>".to_string(), "</u>".to_string());
-    tag_pairs.insert("{i}".to_string(), "{/i}".to_string());
-    tag_pairs.insert("{b}".to_string(), "{/b}".to_string());
-    tag_pairs.insert("{u}".to_string(), "{/u}".to_string());
-    
-    // Identify tag pairs in the original text to preserve tag nesting
-    let mut tag_stack = Vec::new();
-    let mut paired_codes = Vec::new();
-    
-    for (start, end, code) in &original_codes {
-        // Check if this is an opening tag
-        if tag_pairs.contains_key(code) {
-            tag_stack.push((*start, *end, code.clone()));
-        } else {
-            // Check if this is a closing tag and we have a matching opening tag
-            if let Some(opening_code) = tag_stack.last().and_then(|(_, _, _open_code)| {
-                // Find the corresponding opening tag for this closing tag
-                tag_pairs.iter()
-                    .find(|(_, close)| close.as_str() == code.as_str())
-                    .map(|(open, _)| open.clone())
-            }) {
-                if let Some(pos) = tag_stack.iter().position(|(_, _, open_code)| open_code == &opening_code) {
-                    let opening = tag_stack.remove(pos);
-                    paired_codes.push((opening, (*start, *end, code.clone())));
-                }
-            }
-        }
-    }
-    
-    // Analyze the original line structure
-    let original_lines: Vec<&str> = original.lines().collect();
-    let translated_lines: Vec<&str> = translated.lines().collect();
-    
-    // If the line structure completely differs, make best effort to insert codes
-    if original_lines.len() != translated_lines.len() {
-        // For single-line translations, try to preserve tag pairs
-        if translated_lines.len() == 1 {
-            let mut result = String::new();
-            
-            // Add all position codes and standalone symbols at the beginning
-            for (_start, _end, code) in &original_codes {
-                if an_codes.is_match(code) || music_symbols.is_match(code) {
-                    result.push_str(code);
-                }
-            }
-            
-            // Add opening tags
-            for ((_, _, open_code), _) in &paired_codes {
-                result.push_str(open_code);
-            }
-            
-            // Add the translated text
-            result.push_str(translated);
-            
-            // Add closing tags in reverse order
-            for (_, (_, _, close_code)) in paired_codes.iter().rev() {
-                result.push_str(close_code);
-            }
-            
-            return result;
-        }
-        
-        // For multi-line translations that don't match the original structure
-        // Simple approach: add all format codes to the beginning of the translation
-        let mut result = String::new();
-        let all_codes: String = original_codes.iter()
-            .map(|(_, _, code)| code.clone())
-            .collect();
-            
-        result.push_str(&all_codes);
-        result.push_str(translated);
-        return result;
-    }
-    
-    // We'll try to match the original line structure with the translated one
-    let mut result = String::new();
-    
-    // Process each line, adding format codes at similar positions
-    for (line_idx, (orig_line, trans_line)) in original_lines.iter().zip(translated_lines.iter()).enumerate() {
-        // Add a newline between lines (except for the first line)
-        if line_idx > 0 {
-            result.push('\n');
-        }
-        
-        // Extract codes from this specific line
-        let line_codes: Vec<_> = original_codes.iter()
-            .filter(|(start, _, _)| {
-                let line_start = original[..original.find(orig_line).unwrap_or(0)].len();
-                let line_end = line_start + orig_line.len();
-                *start >= line_start && *start < line_end
-            })
-            .collect();
-        
-        if line_codes.is_empty() {
-            // No codes on this line
-            result.push_str(trans_line);
-        } else {
-            // Add codes at relatively similar positions
-            let mut line_result = String::from(*trans_line);
-            
-            // Position calculation will be relative to line length
-            let orig_length = orig_line.len() as f64;
-            let trans_length = trans_line.len() as f64;
-            
-            // Add all codes at the beginning for simplicity if the line is very short
-            if trans_length < 3.0 || orig_length < 3.0 {
-                for (_, _, code) in line_codes.iter().rev() {
-                    line_result.insert_str(0, code);
-                }
-            } else {
-                // Sort tags by type: position tags first, then opening/closing pairs, then others
-                
-                // Step 1: Handle position tags ({\anX}) - always at the beginning of the line
-                let position_tags: Vec<_> = line_codes.iter()
-                    .filter(|(_, _, code)| an_codes.is_match(code))
-                    .collect();
-                
-                for (_, _, code) in position_tags {
-                    line_result.insert_str(0, code);
-                }
-                
-                // Step 2: Handle paired tags (formatting tags that have opening/closing pairs)
-                // Find pairs of tags on this line
-                let line_start = original[..original.find(orig_line).unwrap_or(0)].len();
-                let line_end = line_start + orig_line.len();
-                
-                let line_pairs: Vec<_> = paired_codes.iter()
-                    .filter(|((start, _, _), (end, _, _))| {
-                        *start >= line_start && *start < line_end && *end >= line_start && *end < line_end
-                    })
-                    .collect();
-                
-                if !line_pairs.is_empty() {
-                    // Simple approach for lines with paired tags:
-                    // Add opening tags at start, closing tags at end
-                    for ((_, _, open_code), _) in &line_pairs {
-                        line_result.insert_str(0, open_code);
-                    }
-                    
-                    for (_, (_, _, close_code)) in line_pairs.iter().rev() {
-                        line_result.push_str(close_code);
-                    }
-                } else {
-                    // Step 3: Handle remaining tags (music symbols, etc.) at proportional positions
-                    let remaining_tags: Vec<_> = line_codes.iter()
-                        .filter(|(_, _, code)| !an_codes.is_match(code))
-                        .collect();
-                    
-                    for (start, _, code) in remaining_tags {
-                        // Calculate original relative position (as percentage of line)
-                        let relative_pos = (*start - line_start) as f64 / orig_length;
-                        
-                        // Calculate new position in translated line
-                        let new_pos = (relative_pos * trans_length).round() as usize;
-                        let new_pos = new_pos.min(line_result.len());
-                        
-                        // Insert the code at the calculated position
-                        line_result.insert_str(new_pos, code);
-                    }
-                }
-            }
-            
-            result.push_str(&line_result);
-        }
-    }
-    
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::app_config::{TranslationCommonConfig, ProviderConfig};
-    
-    fn get_test_config() -> TranslationConfig {
-        TranslationConfig {
-            provider: ConfigTranslationProvider::Ollama,
-            common: TranslationCommonConfig {
-                system_prompt: "You are a translator. Translate the following text from {source_language} to {target_language}. Only return the translated text.".to_string(),
-                rate_limit_delay_ms: 0,
-                retry_count: 3,
-                retry_backoff_ms: 1000,
-                temperature: 0.3,
-            },
-            available_providers: vec![
-                ProviderConfig {
-                    provider_type: "ollama".to_string(),
-                    model: "llama2".to_string(),
-                    api_key: "".to_string(),
-                    endpoint: "http://localhost:11434".to_string(),
-                    concurrent_requests: 4,
-                    max_chars_per_request: 1000,
-                    timeout_secs: 30,
-                },
-                ProviderConfig {
-                    provider_type: "openai".to_string(),
-                    model: "gpt-3.5-turbo".to_string(),
-                    api_key: "test-api-key".to_string(),
-                    endpoint: "".to_string(),
-                    concurrent_requests: 4,
-                    max_chars_per_request: 4000,
-                    timeout_secs: 30,
-                },
-                ProviderConfig {
-                    provider_type: "anthropic".to_string(),
-                    model: "claude-3-haiku-20240307".to_string(),
-                    api_key: "test-api-key".to_string(),
-                    endpoint: "".to_string(),
-                    concurrent_requests: 4,
-                    max_chars_per_request: 4000,
-                    timeout_secs: 30,
-                },
-            ],
-        }
-    }
-    
-    #[test]
-    fn test_translation_service_creation() {
-        let config = get_test_config();
-        let service = TranslationService::new(config);
-        assert!(service.is_ok());
-    }
-    
-    #[test]
-    fn test_add_token_usage() {
-        let mut stats = TokenUsageStats::new();
-        stats.add_token_usage(Some(100), Some(50));
-        assert_eq!(stats.prompt_tokens, 100);
-        assert_eq!(stats.completion_tokens, 50);
-        assert_eq!(stats.total_tokens, 150);
-        
-        // Add more tokens
-        stats.add_token_usage(Some(200), Some(100));
-        assert_eq!(stats.prompt_tokens, 300);
-        assert_eq!(stats.completion_tokens, 150);
-        assert_eq!(stats.total_tokens, 450);
-        
-        // Handle None values
-        stats.add_token_usage(None, Some(50));
-        assert_eq!(stats.prompt_tokens, 300);
-        assert_eq!(stats.completion_tokens, 200);
-        assert_eq!(stats.total_tokens, 500);
-        
-        stats.add_token_usage(Some(100), None);
-        assert_eq!(stats.prompt_tokens, 400);
-        assert_eq!(stats.completion_tokens, 200);
-        assert_eq!(stats.total_tokens, 600);
-    }
-    
-    #[tokio::test]
-    async fn test_translate_batches_processes_all_chunks() -> Result<()> {
-        // Create mock data - 5 batches with 2 entries each
-        let batches = vec![
-            vec![
-                SubtitleEntry::new(1, 0, 1000, "First subtitle batch 1".to_string()),
-                SubtitleEntry::new(2, 1001, 2000, "Second subtitle batch 1".to_string()),
-            ],
-            vec![
-                SubtitleEntry::new(3, 2001, 3000, "First subtitle batch 2".to_string()),
-                SubtitleEntry::new(4, 3001, 4000, "Second subtitle batch 2".to_string()),
-            ],
-            vec![
-                SubtitleEntry::new(5, 4001, 5000, "First subtitle batch 3".to_string()),
-                SubtitleEntry::new(6, 5001, 6000, "Second subtitle batch 3".to_string()),
-            ],
-            vec![
-                SubtitleEntry::new(7, 6001, 7000, "First subtitle batch 4".to_string()),
-                SubtitleEntry::new(8, 7001, 8000, "Second subtitle batch 4".to_string()),
-            ],
-            vec![
-                SubtitleEntry::new(9, 8001, 9000, "First subtitle batch 5".to_string()),
-                SubtitleEntry::new(10, 9001, 10000, "Second subtitle batch 5".to_string()),
-            ],
-        ];
-        
-        // Create a progress counter to track callback execution
-        let progress_count = Arc::new(AtomicUsize::new(0));
-        let progress_clone = Arc::clone(&progress_count);
-        
-        // Create a collection to store all processed entries
-        let all_processed_entries = Arc::new(StdMutex::new(Vec::new()));
-        
-        // Process each batch sequentially, simulating the behavior we want to test
-        for (_i, batch) in batches.iter().enumerate() {
-            let processed_entries: Vec<SubtitleEntry> = batch.iter()
-                .map(|entry| {
-                    let mut new_entry = entry.clone();
-                    new_entry.text = format!("[TRANSLATED] {}", entry.text);
-                    new_entry
-                })
-                .collect();
-            
-            // Store the processed entries
-            let mut entries = all_processed_entries.lock().unwrap();
-            entries.extend(processed_entries);
-            
-            // Update progress
-            progress_clone.fetch_add(1, Ordering::SeqCst);
-        }
-        
-        // Verify we processed all batches
-        assert_eq!(progress_count.load(Ordering::SeqCst), batches.len());
-        
-        // Verify all entries were collected
-        let all_entries = all_processed_entries.lock().unwrap();
-        assert_eq!(all_entries.len(), 10, "Should have 10 translated entries total");
-        
-        // Verify all entries have the [TRANSLATED] prefix
-        for entry in all_entries.iter() {
-            assert!(entry.text.starts_with("[TRANSLATED]"), 
-                   "Entry should have [TRANSLATED] prefix: {}", entry.text);
-        }
-        
-        Ok(())
-    }
-}
